refactor: rename config querymodule to configs for consistency.
