+++ lib/model/package.js
@@ -49,7 +49,7 @@ injector.withDefaults = (base, { queries, instances } = {}) => {
     actors: require('./query/actors'),
     all: require('./query/all'),
     audits: require('./query/audits'),
-    config: require('./query/config'),
+    configs: require('./query/configs'),
     fieldKeys: require('./query/field-keys'),
     forms: require('./query/forms'),
     grants: require('./query/grants'),
